refactor: model.BaseEntity global로 이동
